requirements.txt
torch
transformers
pandas
sklearn
python-dotenv
fastapi
uvicorn
starlette
pydantic
base64
plantuml
